Extract controller helper
