+++ app/controllers/spree/admin/payments_controller_decorator.rb
@@ -1,0 +1,3 @@
+Spree::Admin::PaymentsController.class_eval do
+  include SpreeKomoju::ControllerHelpers
+end
