+++ lib/spree_komoju.rb
@@ -1,6 +1,7 @@
 require 'spree_core'
 require 'spree_komoju/engine'
 require 'spree_komoju/errors'
+require 'spree_komoju/controller_helpers'
 
 # This extension adds HTTP PATCH to ssl_request.
 # This is needed for WebMoney multi-card in Komoju.
